refactor: merge manager and interactive_manager
